bcftbx.utils.collections: new module with specialized containers.
Creates a new module for specialized container classes and relocates 'AttributeDictionary' and 'OrderedDictionary' into it.
